Move App component to components folder
